claude/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.40.0
accelerate>=0.27.0
diffusers>=0.27.0
gradio>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
rarfile>=4.1
qwen-vl-utils>=0.0.2
PyMuPDF>=1.24.0